refactor(core): fold finance-api into core-api (ADR-0002 pilot) [REL-14]
Finance becomes the first local widget source served in-process by core
instead of a proxied channel service. Route shapes, cache keys, and
response bodies are byte-identical; the desktop app needs no changes.
The Rust ingester (finance-service) is untouched.

- api/core/sources.go: the localSource seam (dashboard/health/
  invalidateUser hooks) + shared ingestion-probe plumbing for the
  coming sports/predictions/rss folds
- api/core/finance.go: /finance routes, dashboard section, health
  probe via INTERNAL_FINANCE_URL, per-user cache invalidation
- server.go/channels.go: local sources aggregate in-process; discovered
  channels with the same name are skipped so a still-registered
  finance-api can't double-report during cutover (native routes already
  win — they register ahead of the dynamic proxy)
- Dropped per REL-10 (ADR-0002 Appendix A): /internal/cdc handler and
  all finance:subscribers:* set writes — every reader was dead
- Removed: channels/finance/api (module, Dockerfile, compose files),
  k8s/finance-api.yaml, its NetworkPolicy pair (finance-service now
  accepts core-api), deploy matrix row + path filter, backend-tests
  entry, smoke-test entry
- INTERNAL_FINANCE_URL moved channels-config -> core-config
- Docs: AGENTS.md, LOCAL_SETUP.md updated

Post-merge manual cleanup (kubectl apply does not prune):
  kubectl delete deployment/finance-api service/finance-api -n scrollr
  kubectl delete networkpolicy allow-core-to-finance-api allow-finance-api-to-finance-service -n scrollr
