--- channels/finance/api/.dockerignore
@@ -1,10 +1,0 @@
-# Local build artifacts + secrets must not enter the image build context;
-# the Go binary is rebuilt inside the image.
-.env
-.env.*
-*.pem
-*.exe
-main
-*-api
-*_api
-scrollr-*
