--- channels/finance/api/Dockerfile
@@ -1,21 +1,0 @@
-FROM golang:1.25-alpine AS builder
-
-WORKDIR /app
-COPY go.mod go.sum ./
-RUN go mod download
-COPY . .
-RUN CGO_ENABLED=0 GOOS=linux go build -o finance-api .
-
-FROM alpine:latest
-
-# Sentry release tagging — the runtime reads GIT_SHA via os.Getenv.
-ARG GIT_SHA=unknown
-ENV GIT_SHA=${GIT_SHA}
-
-RUN apk --no-cache add ca-certificates
-WORKDIR /app
-COPY --from=builder /app/finance-api .
-
-EXPOSE 8081
-
-CMD ["./finance-api"]
