--- channels/finance/docker-compose.dev.yml
@@ -1,28 +1,0 @@
-version: "3.8"
-
-services:
-  scrollr-finance-api:
-    build:
-      context: ./api
-      dockerfile: Dockerfile
-    environment:
-      - DATABASE_URL=${DATABASE_URL}
-      - REDIS_URL=${REDIS_URL}
-      - ALLOWED_ORIGINS=${ALLOWED_ORIGINS}
-      - CHANNEL_URL=https://finance.api.myscrollr.enanimate.dev
-      - INTERNAL_FINANCE_URL=http://scrollr-finance-service:3001
-    depends_on:
-      scrollr-finance-service:
-        condition: service_started
-    restart: unless-stopped
-
-  scrollr-finance-service:
-    build:
-      context: ./service
-      dockerfile: Dockerfile
-    environment:
-      - DATABASE_URL=${DATABASE_URL}
-      - TWELVEDATA_API_KEY=${TWELVEDATA_API_KEY:-test_twelvedata_key}
-      - TWELVEDATA_REST_URL=http://mock-twelvedata:9002
-      - TWELVEDATA_WS_URL=ws://mock-twelvedata:9003/v1/quotes/price
-    restart: unless-stopped
