--- channels/finance/docker-compose.yml
@@ -1,35 +1,0 @@
-version: "3.8"
-
-services:
-  scrollr-finance-api:
-    build:
-      context: ./api
-      dockerfile: Dockerfile
-    container_name: scrollr-finance-api
-    # Bind only to localhost — the core gateway is the only trusted caller
-    # and it reaches us via the Docker network (CHANNEL_URL), not the
-    # host-published port. Publishing on 0.0.0.0 let anything else on the
-    # host bypass the gateway and hit our routes directly (channels don't
-    # validate JWTs — they trust the X-User-Sub header set by the gateway).
-    ports:
-      - "127.0.0.1:8081:8081"
-    environment:
-      - DATABASE_URL=${DATABASE_URL}
-      - REDIS_URL=${REDIS_URL}
-      - ENCRYPTION_KEY=${ENCRYPTION_KEY}
-      - ALLOWED_ORIGINS=${ALLOWED_ORIGINS}
-      - CHANNEL_URL=${CHANNEL_URL}
-      - INTERNAL_FINANCE_URL=http://scrollr-finance-service:3001
-    restart: unless-stopped
-
-  scrollr-finance-service:
-    build:
-      context: ./service
-      dockerfile: Dockerfile
-    container_name: scrollr-finance-service
-    ports:
-      - "3001:3001"
-    environment:
-      - DATABASE_URL=${DATABASE_URL}
-      - TWELVEDATA_API_KEY=${TWELVEDATA_API_KEY}
-    restart: unless-stopped
