+++ k8s/configmap-channels.yaml
@@ -9,14 +9,14 @@ data:
   FRONTEND_URL: "https://myscrollr.com"
 
   # Channel API self-registration URLs (K8s Service DNS)
-  FINANCE_CHANNEL_URL: "http://finance-api:8081"
+  # (finance-api folded into core-api — ADR-0002, REL-14)
   SPORTS_CHANNEL_URL: "http://sports-api:8082"
   RSS_CHANNEL_URL: "http://rss-api:8083"
   FANTASY_CHANNEL_URL: "http://fantasy-api:8084"
   PREDICTIONS_CHANNEL_URL: "http://predictions-api:8085"
 
   # Go API -> Rust service internal URLs (K8s Service DNS)
-  INTERNAL_FINANCE_URL: "http://finance-service:3001"
+  # (INTERNAL_FINANCE_URL moved to core-config — ADR-0002, REL-14)
   INTERNAL_SPORTS_URL: "http://sports-service:3002"
   INTERNAL_RSS_URL: "http://rss-service:3004"
   INTERNAL_PREDICTIONS_URL: "http://predictions-service:3005"
