+++ k8s/configmap-core.yaml
@@ -20,6 +20,10 @@ data:
   ALLOWED_ORIGINS: "https://myscrollr.com,https://api.myscrollr.com"
   FRONTEND_URL: "https://myscrollr.com"
 
+  # Rust ingestion services core probes directly (local widget sources,
+  # ADR-0002). Moved from channels-config as each Go channel API folds in.
+  INTERNAL_FINANCE_URL: "http://finance-service:3001"
+
   # Logto role IDs (not secret, just config)
   LOGTO_UPLINK_ROLE_ID: "x8qz9zmdthw19n10dqnr5"
   LOGTO_PRO_ROLE_ID: "mpp4isevqg194srzrhjha"
