+++ scripts/smoke/production-readiness.sh
@@ -79,7 +79,6 @@ SERVICES=(
     "finance-service 13001  3001  /health/ready"
     "rss-service     13004  3004  /health/ready"
     "sports-api      18082  8082  /internal/health"
-    "finance-api     18081  8081  /internal/health"
     "rss-api         18083  8083  /internal/health"
     "fantasy-api     18084  8084  /internal/health"
     "predictions-service 13005 3005 /health/ready"
